requirements.txt
# --- API & Database ---
alpaca-py==0.31.0
supabase==2.11.0
python-dotenv==1.0.1

# --- Technical Analysis & Data ---
pandas==2.2.2
numpy==1.26.4
ta==0.11.0

# --- Orchestration & Scheduling ---
schedule==1.2.2